+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## 0.6.0 - 2025-10-06
+
 - Add `:prompt()` text decorator.
 - Notify when trying to send to a terminal that hasn't been started yet.
 - Available text decorators by name can be configured in `config.text_decorators`.